requirements.txt
python-telegram-bot==13.7
anthropic==0.3.6
python-dotenv==0.19.2
pydantic>=1.9.0,<2.0
